requirements.txt
pytesseract
pyautogui
pynput
opencv-python
pillow
pydirectinput
tensorflow-gpu
numpy
pymongo
dnspython
